set connection properties in JetpackCS
While looking through the telecom dumpsys logs, I noticed there was
multiple swtiches in AUDIO_ROUTES before an outgoing call would go
active. Upon investigation this was due to not setting the connection to
dialing in the JetpackConnectionService.  The
ConnectionServiceFocusManager would kick the ongoing call off as the
focus call because the call state was in a non-focusable state (NEW).

Additionally, the folllowing Connection properties are now being set in
JetpackConnectionService:
- displayName
- address
- next callState (dialing or ringing )
- callType / video state
- capabilities

And, I've removed the CALL_PHONE permission.

bug: 279043628
Test: JetpackConnectionServiceTest class
Change-Id: I077ffdb387a087b95b444943cd1d5dc1c976d35b