Add rectangle shape function
This library was created with the original idea of making round rects
better, so it would be good to offer an easy way to get such shapes.
It is not hard to get a rectangular shape (rounded or not) from the
current APIs, but it is requires creating a vertex list since the
easier RoundedPolygon constructors create shapes with edges of
equal sizes.

This commit also changes the default number of vertices in circle()
to 8, from the previous 4, just to up the quality of the result for
this common shape.

The Compose and View integration tests were updated to include
examples of these shapes.

Bug: 280322189
Test: existing and new unit tests pass, manual testing with integration tests
RelNote: Added new RoundedPolygon.rectangle() function
Change-Id: I78e7ee609295c20680c5e6f1a3383421420674ff