Enable WIC#show(ime()) compat mode also for API 34+
This is a follow up CL to our recent CL [1], which aimed to fill the
remaining behavior gap between

  WindowInsetsController#hide(ime())

and

  InputMethodManager#hideSoftInputFromWindow()

in multi-windowing mode (e.g. Bubble and split-screen) devices by
falling back to IMM#hideSoftInputFromWindow() on pre-U devices.

Unfortunately the fallback seems to be still necessary on recent U
devices.  Let's remove the version gating for now to provide a
workable solution for app developers.

 [1]: I1db12657b9ec6f6cc7406e7a729a43d7acd6fada
      14c150a576cc014691fac15d0011725de59e7872

Bug: 258048232
Bug: 280532442
Fix: 278321935
Test: ImeMultiWindowTest
Change-Id: I82895a8cab81e2543e448437a10418ac02432d26