Fix FutureGarbageCollectedException issue caused by BasicExtenderSessionProcessorTest
BasicExtenderSessionProcessorTest should call deInitSession() to close the used DeferrableSurface list when the tests have been executed completely. So that the completer object of the DeferrableSurface won't be gargabe collected to cause the FutureGarbageCollectedException.

Although the issue was reported for the ExtraSupportedSurfaceCombinationsContainerDeviceTest, the log contains "CameraX-AutoDrainThread" which is the string used to create the HandlerThread in BasicExtenderSessionProcessorTest.

Bug: 278859779
Test: BasicExtenderSessionProcessorTest
Change-Id: I7ef0f7ef0d7701c7df027bd7b60aac262e3d3a61