Remove 5s delay in FocusMeteringDeviceTest and ZoomControlDeviceTest
In FocusMeteringDeviceTest and ZoomControlDeviceTest, a delay of 5s was
added during shutdown because of b/264396089 where CameraPipe wouldn't
close the camera if an `ActiveCamera` was never connected. Now that the
bug is fixed, we can actually just remove these delays and closing down
all camera devices (b/264332446) isn't really necessary here.

Bug: 264396089
Test: Run FocusMeteringDeviceTest and ZoomControlDeviceTest on CameraX's
  suggested devices (API levels ranging from 23 to 33) and passes both
  tests 3/3 times.
Change-Id: I2f4b616ba447534b9b23c8033090a6ab437d9341