Move to only building APKs that are required for testing
Before this change we built and copied both debug and release
build types for test apps even though most of the time we only
need debug (except for benchmarks that use release apk)

This reduces the number of tasks run for zipTestConfigsWithApks
from 42,403 to 34,348. It also makes androidTest.zip file go from
4.9G to 2.9G in size. Performance to be seen when we land this
in production, but it should help androidx_device_tests target.

Bug: 280632101
Test: ./gradlew zipTestConfigsWithApks
Change-Id: Id4c4a2622c4db902e6b5e611cee8113bcbd072e0