Unify getLineVisibleEnd results between API levels
This CL attempts to remove the `getLineVisibleEnd` calculation difference between APIs below and above level 26. There was a major change in API 26 that affected how total line count is calculated in StaticLayout when maxLines argument is present. This has been addressed in Compose for `TextLayout.lineCount` but other APIs that depend on this value have not been unified.

Here, we are making sure that lineVisibleEnd returns the same result for all API levels by calculating it without checking for lineCount.

Unfortunately there is a bigger difference below API 23 that affects how text line bounds are calculated. We cannot unify the lineVisibleEnd since lines themselves are drastically different. Hence, there are some SDK version checks in some tests.

Fix: b/270441925
Test: :compose:foundation:foundation:cAT
Test: :text:text:cAT
Change-Id: Id92ac882d31ec99760292cc27bafbf3f8b4a7db7