[CameraPipe] Add pending and resubmitting mechanism for still capture requests
Still capture requests may be discarded if they are issued for a UseCaseCamera where CameraGraph has not started yet and the UseCaseCamera is recreated afterwards. With this change, we will add such requests to a pending list and retry them again when a new UseCaseCamera is set.

The requests can be added to the pending list in two cases.
1. Request is not submitted to CameraGraph due to current CameraGraph is in closing/closed state.
2. Request is submitted to CameraGraph but aborted by CameraGraph instead of submitting to framework, usually due to the closing of CameraGraph.

Bug: 280560222
Test: ImageCaptureTest, StillCaptureRequestControlTest, ImageCaptureSwitchCameraStressTest
Change-Id: Id4f76ebd31adb1dddc94626b67d17c5f370f42d6