Provide Fragment Predictive back support
Providing support for the new AndroidU Predictive back APIs in
Fragments. This change makes it so that using gesture back with
Animators in Fragments is now seekable meaning you can see if the
previous destination without completely going back to it.

RelNote: "Fragments now provide support for Predictive back when using
  Animators. This allows you to use the back gesture motion to seek to the
  previous fragment with your custom Animator before deciding to either
  commit or cancel the transaction via the completed gesture."
Bug: 258863498
Test: Added new back stack change listener tests and sample app

Change-Id: I36c582d0d8e8b44eb0ee5cf2db9498271cf49211