Return oversized BIV if no BIV is within the viewport size
When finding the next BIV request, if there is a request
that is oversized, the algorithm returns the largest BIV
that fit in the viewport.
However, if there are no BIV requests smaller than the
viewport, the oversized BIVs are ignored, causing scroll
to fail.
Updating code to return the smallest oversized BIV request
if all requests are oversized.

Test: Fixes failing test.
Fixes: 278219642

Change-Id: I53af354b01b82eaba246ec4e96cc61e93ed20eb5