Add windowInputToken parameter to SandboxedUiAdapter
This input token is used when creating the SurfaceControlViewHost
that will host the UI provider's content. This is necessary to
ensure that the SurfaceControlViewHost's token is correctly
associated with the host of the embedded hierachy.

The input token is obtained by creating a throwaway SurfaceView whenever
the SandboxedSdkView attaches to a window. The input token will be obtained
when the SurfaceView is attached. The SurfaceView will be removed immediately
after it has been attached to the window. The throwaway SurfaceView
is never visible.

Bug: 283748196
Test: Manual, SandboxedSdkViewTest
Relnote: Added windowInputToken parameter to SandboxedUiAdapter
Change-Id: Ief578bfa0202f8577e6a426c5b049f763c0b4846