Implement service to client notification of isolate termination
Allows the client side to provide the service with an interface for
calling into the client side, and use this mechanism to receive
notifications when the service side has terminated due to a crash.

MemoryLimitExceededException is no longer expected to throw for
exactly one evaluation. This was never strictly guaranteed by the
service to begin with. The service is being updated to signal zero,
one, or many memory limit exceeded errors via callbacks. Public docs
did not previously guarantee this. Out of memory conditions will now
primarily be communicated via the new termination notification
mechanism.

Bug: 261442164
Bug: 261693855
Test: ./gradlew :javascriptengine:javascriptengine:connectedAndroidTest --info
Change-Id: Icf359de9cc83d63b376bf77b31c7873b04ba59dc
Relnote: Make MemoryLimitExceededException and SandboxDeadException subclass IsolateTerminatedException