Transfer touch focus to host when onScroll/onFling detected
Adds a gesture detector to the remote view created by the UI
provider. If an onScroll or onFling gesture is detected, the touch
focus is transferred to the host. For all other gestures, nothing
is intercepted.

The touch interception happens on a UI library owned ViewGroup which
is a parent of the UI provider's view. This ensures that the SDK's
onTouchListener is not overridden.

Relnote: Added logic to transfer touch focus to host
Test: Manual
Test: BinderAdapterDelegateTest
Bug: 283748196
Change-Id: I33b54b772e2802e2f987aa2414e06f83bafd9e7f