Stop closing recomposer during the benchmark runs.
`disposeContent` is intended to run repeatedly during the benchmark, e.g. when measuring first composition time. It, however, calls `recomposer.close`, which is a destructive operation which affects subsequent recompositions. After content is disposed, `recomposer` is not recreated, making benchmark potentially incorrect.
As removing view from the hierarchy disposes its composition, we can rely on view removal to do content disposal instead of using recomposer for that.

This change extract `recomposer.close` into a separate method which is only called after benchmark was completed.

Some benchmark results might be degraded by this change, which is expected, as the results will be more correct.

Change-Id: Iaec9983136a200d01332d18c92e19ef82aaacd5a