Fix Dialog not marked complete
When multiple dialogs are open at the same time and the top one is dismissed, the second dialog in backstack is marked as transitioning to hold it in STARTED state until the top dialog has completed transitioning. But the second dialog was never marked as complete so it stays in STARTED state instead of moving to expected RESUMED state.

Now we make sure to mark Dialogs as complete when Dialog has been commposed.

Test: ./gradlew navigation:navigation-compose:cC
Bug: 286371387
Change-Id: I48da82d4c0db3fecd40b4357b152e807078066c6