Update BTF2 filter through restart session code path.
I don't think there's any reason to have a separate code path for
setting the filter on a live session – it's simpler to just restart the
session from scratch when the filter changes. I don't think a lot of
people will be dynamically changing the filter parameter during an edit
session anyway.

Test: no changes
Change-Id: If4e4367f65b29fa18cecb534396b7dd248c640d8