Update ConstraintLayout documentation
Elaborates on usage with examples.

For both the version with the inline Modifier DSL and the ConstraintSet
DSL.

The ConstraintSet DSL is thororughly documented on the ConstraintSet
interface, since that's where kotlin resolves to in the IDE for KDoc
(Javadoc equivalent in Kotlin). And it doesn't allow for referencing
specific overloads of the same reference.

The KDoc in the ConstraintLayout Composable that takes ConstraintSet
instead focuses on good practices.

Bug: n/a
Test: n/a
Change-Id: I496c99cf1fceb9efd7c091a4592cb4a05bcbfb8e