Add test coverage for not setting callbacks
Add test coverage to verify that a CallException is thrown when the
client does not first set the call control callback. Ensure that in the
legacy code path that we always send the disconnect signal before
throwing the exception so that we do not end up with a stuck call.

Fixes: 284993816
Test: BasicCallControlCallbacksTest
Change-Id: Ic2d433279cc2d48d65933b44eccceef00d78ed06