Replace stately with internal atomics
When KMPfying a Jetpack project, it is important to ensure
the Android targets are not negatively affected, even if it means
more duplicated code between common and android targets.

Previous implementation added a new dependency (stately) which is not
required for Android.

This CL replaces it with internal implementations which were already
available in collections.

A followup project needs to create some common artifacts for Jetpack
KMP internals but for now, copying is the only option to ensure KMP
changes do not affect android.

I've also fixed a bug in paging on how we were using copy on write list
wrong (unrelated to stately, but it was still broken with stately). Added
a new test for it.

Bug: n/a
Test: InvalidatingPagingSourceFactoryTest
Test: ANDROIDX_PROJECTS=INFRAROGUE gw :paging:paging-common:allTests
Change-Id: Ic37f0e3f0bb8ff31795f6837da8714fbde383aaf