Update rear display mode API's
Removes session-based API for rear display mode
to better reflect that the entire device
is moving to the rear facing display, as well
as how it is sticky even if the app that requests
the move is backgrounded or killed.

This CL also updates the presentContentOnWindowArea method
to take a WindowAreaToken object instead of a Binder token

Moves the `WindowAreaInfo#getActiveSession` API to be
more correctly on the `WindowAreaController`.

This is a breaking CL in the attempt to stabilize and remove
the experimental tag on the Window Area API's.

Relnote: "Renamed transferActivityToWindowArea to transferToWindowArea
 to be clearer about the behavior when triggering the feature. This also
 removes the session based behavior for this method as it's an operation
 that outlives the process that requests it."
Bug: 271644150
Fixes: 273807246
Fixes: 272064992
Fixes: 473870239
Test: WindowAreaControllerImplTest
Change-Id: Ie7d6fa7258d323363748e25fade8d658c1f0f510