Cleanup TestIO to be more consistent
This CL updates the TestIO abstraction that we use in Datastore.

I pushed down most of the implementation to the File/Path wrappers
such that the chances of diverging is minimal.

I've also added checks to ensure any fresh TestIO instance starts
with a new fresh folder (that is check created in initialization).

Also aded some simple tests to ensure they stay consistent with
eachother.

Fixes: 276983736
Test: ANDROIDX_PROJECTS=INFRAROGUE gw test allTest connectedChecks

Change-Id: I8e732c78c5bf4a28b29518d9c971f59c629396d1