Fix expandedItem's behavior when they have buttons.
If a expandedItems has buttons, we need to set the z-index of the collapsed
and expanded versions of the content so iuser interactions matches with
what the user sees.
I.e. if we are seeing the expanded version, the expanded controls are
clickable.

Relnote: "Fix expandedItem's behavior when they have buttons."
Test: Added
Bug: 289991514

Set the z-index of expandables so the

Change-Id: I1899d8055e5ebd89263fc498964565215452e95c