Resolving potential memory leak vulnerability in QueryInterceptorStatement.
Each time a `bind*` in QueryInterceptorStatement is invoked, `saveArgsToCache` is called to save the new argument being processed into the `bindArgsCache`.

This works as intended for all `bind*` functions except `bindNull()`, due to the fact that the `bindArgsCache` itself is is passed in for the `value` parameter of `saveArgsToCache` instead of a `null` value. This results in the entire cache being placed at the given index within `bindArgsCache` as opposed a single null value.

Bug: 290641662
Test: QueryInterceptorTest.kt
Change-Id: If6c711f836ac1e90aed55bccdd24bcbd9e36bf6f