Defer AndroidView update callback until view is attached.
The snapshot observer is only started when the view is attached, so if
we run the update callback before that, none of the state reads will be
observed.

This bug is not triggered in a test if the state is set by the test –
it's only triggered if the state is updated by an effect. Added a test
that does that and fails before this fix.

This is also technically more efficient, since it avoids running the
update callback if the view never becomes attached. I think it also
means we can use the Owner's SnapshotStateObserver instead of creating
our own, although I haven't made that change in this CL.

Fixes: b/291094055
Test: AndroidViewTest.androidView_updateObservesStateChanges_fromEffect
Relnote: "`AndroidView`'s `update` callback's first invocation will now
    be defered until the view is attached, instead of running when the
    composition that introduces the `AndroidView` is applied. This fixes
    a bug where the `update` callback wouldn't be invalidated if a state
    it read was changed immediately by an effect."
Change-Id: Ie94381c17c417b4161d1531c9731a3b077ad46fb