Fix issue with KSTypeVarianceResolver for typealiases.
This CL fixes an issue where we were getting an
"IndexOutOfBoundsException" for cases where the type arguments of the
actual type are a different size than the type arguments of the
originally declared type. This mismatch in size was due to not replacing
typealiases in the originally declared type, thus it causes issues if
the typealias has a different number of type args than the type its
aliasing, e.g. `MyAlias<T> = MyClass<T, T>`.

Test: KspTypeNamesGoldenTest.kt
Change-Id: I363c0493142dc5005e920f0cc6872cf0c7cdf1e6