Improve accessibility pseudo-scroll-offsets with Lazy lists
The previous math did not take into account the fact that this value
winds up cast to an integer in the AccessibilityEvent, so in practice
the firstVisibleItemScrollOffset was rounded away. This was observed to
degrade TalkBack focus behavior during two-finger scrolling.

This patch replaces it with different math which is less likely to be
monotonically increasing (monotonicity is not known to matter for any
real behavior anyway), but has a much stronger likelihood of surviving
the series of casts (first to Float, and then to Int).

Related minor cleanups:
1. Refactor the math into helper methods on the interface.
2. Simplify the pseudo-maxScrollOffset math to be always be a fixed
integer above the current scroll.
3. Delete the `setScrollEventDelta` call on the AccessibilityEvent,
because the call does not have a known benefit and these deltas
were far too noisy to be usable for anything.

Bug: 286202600
Test: LazyListScroll.animateScrollBySemantics

Change-Id: I158781cf2cf4f8824d281e91b887d25d6ee86df2