make addCall a blocking function
It was observed that when wrapping addCall in a try/finally,
the finally block would be entered as soon as the addCall session
started.  This made catching exceptions and doing cleanup extremely
hard.

In order to use try/catch/finally blocks, addCall needs to be blocking
while the session is alive.  To acomplish this, a CompletableDeffered
awaits for a disconnect in order to complete the session.

Test:  BasicCallControlsTest#testTiming
Fixes: 291604411
Change-Id: I385482a91e68759deff9a8f07ba49eacc1aceb14