addCall docs update for 5 sec notif window
outgoing calls that fail to post a notif within 5 seconds will lose
foreground service delegation.  This bug was observed in the
public telecom connectivity test app. The outgoing call was left in the
dialing state for over 5 seconds after adding the call to Telecom via
addCall. This lead to the app losing FGS delegation even though a notif
was posted after 5 secs.

Fixes: 289544518
Test: none
Change-Id: I855391d2c84dedec2e1f9fc705f898423a4126f3