Suppress deprecated icons usage in the TV libs
This CL suppresses a deprecated icons usage for icons that should be
auto-mirrored on RTL layouts.
Note that we opted to suppress the warning in your library to
maintain the current behavior, but please consider updating your code by
unpinning the Material icons or extended-icons libraries, or by pinning
them to the latest version after the next release.
Once you have done so, you'll be able to use the updated icons under the
new `automirrored` package that the deprecation message suggests to use.

Test: N/A (deprecation suppression only)
Relnote: N/A
Bug: 279989215
Change-Id: Ib45c4025958325377b46898a322959b10f20c0cf