Guard against missing anchor ids
With compose I see some cases where we don't receive any anchorId
this has been seen with compose.ui:ui version 1.4.3

It looks like CompositionGroup.identity would have to return null
for this to happen. It is unclear why this is happening.

This change adds a safeguard for this.
If the anchorId is not available: recreate the entire tree and find
the compose node we need parameters for.

Bug: 291789498
Test: Updated existing tests
Change-Id: Ic0bb21795f0164284fd625ac2d3749c1d0a8a54d