Move compose inspector setup from Wrapper to the inspector
This will ensure that the slot table is available in each
AndroidComposeView when needed by the compose inspector.

Before we relied on debug view attribute setting to be on, and
when it was detected any new AndroidComposeView would be tagged
with a slot table which would be populated by the runtime.

There are 2 problems with this approach:
- The LI would set the debug view global attribute which caused
  an activity restart which in turn would add the slot table.
  We have reported cases where the LI would ask for composables
  before this mechanism was ready to populate the slot table.
- If the view debug flag was on without the Layout Inspector
  being attached, the runtime would unneccessarily create slot
  tables and gather compose information.

Now this is done when the compose inspector is loaded.
- First we register a hook such that we can add a slot table to
  new AndroidComposeViews being created.
- Then we add slot tables to all existing AndroidComposeViews.
- If any slot tables were added we perform a hot reload to
  populate the tables.

Fixes: 288152418
Test: Modified existing tests
Change-Id: I7ab73c11fe4a2a187e5c4d55dc7cfa0732687856