Safeparcelables in AppSearch (1) - Include SafeParcelable
Copy SafeParcelable classes with modifications in AppSearch. We will keep this copy and update them rarely to avoid including the entire Safeparelable in Jetpack, in which it is not needed.

classes duplicated:
- SafeParcelable
- AbstractSafeParcelable

Modifications are done to remove real implementation but keep the
annotations for code sync purpose with framework.

Annotation processor is not included. So for CREATOR classes, we create stubs and use them in Safeparcelable classes.

Bug: 275629842
Test: ./gradlew appsearch:appsearch:connectedCheck
Change-Id: I86c97ae6e5e6a194a3d66076667bff777cd348c0