Add timeouts to BasicCallControlsTest tests.
Adding 10s timeouts to the BasicCallControlsTest to prevent tests from
hanging. There was a test failure in testIsMuted_BackwardsCompat which
caused the test to hang for 45 minutes (likely due to waiting for the
mute state change to be received in the nested coroutine). All tests are
for the most part upperbounded by a ~5 second execution time so 10s
should be ample.

Bug: 290699346
Test: BasicCallControlsTest
Change-Id: I7efdb4d0ac02003a55acab9394fc227d5784a916