Bugfix: Allow PendingValue to be set again.
In CameraController, we store the value of CameraControl if the camera is not yet initialized. e.g. zoom ratio. Then, we propagate the value after the  initialization. However we made an incorrect assumption that the camera is only initialized once. So we throw Exception if the future is completed.

In fact, the camera can be intialized twice when the app calls unbind. The following code throws an unexpected Exception:

```
cameraController.setZoomRatio(1)
cameraController.unbind()
cameraController.setZoomRatio(1)
```

The fix in this CL removes the check for Future completion and make sure that each value is only propagated once.

Bug: 294000691
Test: manual test and ./gradlew bOS
Change-Id: Ib5773bb22f89021694ba96c7b49536c7ec97bcf0