Fix CC invalidation when switching between IDE and cmdline
Set org.gradle.projectcachedir in gradle.properties so that Studio
picks up this value when running Gradle commands. We set this value
in gradlew, but Studio does not run gradlew, so gradle.properties
fixes that. It does have a short-coming that OUT_DIR being set is
not supported within studio, but it is way better than what we had
before. The main reason this is important is because kotlin gradle
compiler puts incremental compilation data into this cache directory
and having it move (prior to this change) was causing buildSrc to
be rebuilt every time when switching between the two.

Bug: 279090332
Test: ./gradlew tasks
      ./studiow all (and gradle sync)
      ./gradlew tasks -> hit the configuration cache
Change-Id: Icfce7061d996d05638d66da082cf8b6e4f8b3301