Fix unplaced child doesn't get placed
Fixes: 257134777

A child that isn't placed due to one of its modifiers
wasn't able to be placed again when that modifier
changed. This fixes that by looking at isPlacedByParent()
rather than isPlaced() and allowing an invalidated
layout to be placed if it never had been placed before
and its parent tried.

Test: new test

Change-Id: I7f313314f69d05419157bfae4447fc22024616e5