Make TextFieldBuffer not implement CharSequence.
There was a lot of feedback in the user study that the extension methods
available on CharSequence were very confusing when working with TFB,
because those methods all return new sequences instead of mutating
in-place.

The easiest way to avoid that problem is to simply not implement
CharSequence. It doesn't have any strong reason to need to implement it
anyway. Instead, I've given it an asCharSequence() method that returns
an implementation of CharSequence without allocating so all the
extensions of CharSequence are still available with just one additional
level of indirection.

Test: TextFieldBufferTest
Relnote: n/a
Change-Id: I1a2ec7b57ce1430e16081db068f97ee0cba09584