Add a lint rule to ensure progress flow is collected
The new PredictiveBackHandler API provides a flow to the onBack lambda
that must be collected in order to ensure the API gets all the correct
callbacks.

We have runtime checks in place, we should add a lint rule as well for
compile time safety.

This rule will warn in the following cases:
1. The given flow is not referenced at all
2. The given flow does not call collect()

RelNote: "Provided a new lint rule that warns when using the
`PredictiveBackHandler` API and not collecting on the Flow."
Test: added lint test
Bug: 294906729

Change-Id: If47a24f34239c203adaa332df3eebc37b0a11262