Wait for the UI thread after BaseRecyclerViewInstrumentationTest.waitForIdleScroll
onScrollStateChanged is not necessarily called after all relevant
main-thread effects have occurred. Tests that call waitForIdleScroll
and then perform assertions are liable to have those assertions flake
due to thread-safety issues (visibility and/or race conditions).

To fix this, we wait for UI thread to finish what it's doing before
continuing.

The particular test this aims to fix is
PoolingContainerRecyclerViewTest#sharedViewPool, but the fix should
hopefully be broadly applicable to other tests using this method.

Test: Ran PoolingContainerRecyclerViewTest#sharedViewPdool, will use
presubmit to verify nothing else broke.
Fixes: 295582399

Change-Id: Ie95f2ab1671d71a63480703f642f5eaefc58c609