Make sure Modifier.pointerInput() uses the latest lambda after reuse
One of the recommended patterns for using Modifier.pointerInput() is:

val currentValue by rememberUpdatedState(value)
Modifier.pointerInput(Unit) {
   // use currentValue
}

However after we migrated this modifier to Modifier.Node this pattern is broken after reuse happens as we do not call update when the modifier element didn't change and the key representing the element identity is Unit, which didn't change. When lambda changes we don't update the modifier. Technically correct would be to recommend our users to write

val currentValueState = rememberUpdatedState(value)
Modifier.pointerInput(currentValueState) {
   // use currentValueState.value
}

instead, however we can't really expect users to write such code as it is overcomplicated and counterintuitive.
This case is quite unique and we don't have any other modifiers like that so we decided to workaround it right in the Modifier.Node system for this type of modifier. After this change the update will be called for this modifier after reuse even when the input didn't change. If in future we will figure out there are other use cases like that we will consider adding a public api to solve the same issue from other implementations.

Fixes: 293820085
Test: SuspendingPointerInputFilterTest
Change-Id: Iee6f3fac87d87106791a33ea887a478df86cfce1